fixes and removed unused imports
